+++ app/Http/Controllers/TasksController.php
@@ -20,6 +20,10 @@ class TasksController extends Controller
      */
     public function index()
     {
+        if (auth()->user()->cannot('viewAny', Task::class)) {
+            abort(403, "You do not have permission to view tasks.");
+        }
+
         return Inertia::render('Tasks/Index', [
             'page' => Task::with('project', 'user', 'client')->paginate(10),
         ]);
@@ -30,6 +34,10 @@ class TasksController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->cannot('create', Task::class)) {
+            abort(403, "You do not have permission to create tasks.");
+        }
+
         return Inertia::render('Tasks/Create', [
             'users' => User::select('id', 'name')->get(),
             'clients' => Client::select('id', 'name')->get(),
@@ -42,6 +50,10 @@ class TasksController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        if (auth()->user()->cannot('create', Task::class)) {
+            abort(403, "Unauthorized action.");
+        }
+
         $task = Task::create($request->all());
         $task->save();
         return redirect('tasks');
@@ -52,6 +64,10 @@ class TasksController extends Controller
      */
     public function show(Task $task)
     {
+        if (auth()->user()->cannot('view', $task)) {
+            abort(403, "You do not have permission to view this task.");
+        }
+
         return redirect("tasks/{$task->id}/edit");
     }
 
@@ -60,6 +76,10 @@ class TasksController extends Controller
      */
     public function edit(Task $task)
     {
+        if (auth()->user()->cannot('edit', $task)) {
+            abort(403, "You do not have permission to edit this task.");
+        }
+
         return Inertia::render('Tasks/Create', [
             'users' => User::select('id', 'name')->get(),
             'clients' => Client::select('id', 'name')->get(),
@@ -73,6 +93,10 @@ class TasksController extends Controller
      */
     public function update(UpdateRequest $request, Task $task)
     {
+        if (auth()->user()->cannot('update', $task)) {
+            abort(403, "Unauthorized action.");
+        }
+
         $data = $request->only(['title', 'description', 'assigned_to', 'for_client', 'related_to_project']);
         $data['status'] = filter_var($request->input('status'), FILTER_VALIDATE_BOOLEAN);
 
@@ -103,6 +127,10 @@ class TasksController extends Controller
      */
     public function destroy(Task $task)
     {
+        if (auth()->user()->cannot('delete', $task)) {
+            abort(403, "Unauthorized action.");
+        }
+
         $task->deleteOrFail();
         return Inertia::render("Tasks/Index");
     }
@@ -112,6 +140,10 @@ class TasksController extends Controller
      */
     public function updateStatus(Request $request, Task $task)
     {
+        if (auth()->user()->cannot('update', $task)) {
+            abort(403, "Unauthorized action.");
+        }
+
         $task->update(['status' => $request->status]);
         return Inertia::render("Tasks/Index");
     }
